Remove 4addr udev rule hack
Remove udev rule that puts wireless devices in 4addr mode. This
rule was a hack to workaround an issue resulting from networkd
trying (and sometimes failing) to add wireless links to the bridge
prior to hostapd starting on boot.

This issue no longer seems to occur with 5.1+ kernels and if it does
re-appear a better solution is to update the hostapd service units
with 'Before=systemd-networkd.service' for force networkd to wait.