Allow network to be 'up' without any bridge slaves
Our only requirement for the network to be deemed up is for the
LAN bridge to have an IP. We should not need to wait for hostapd
APs to be up or for wired connections to be plugged in.

Set 'ConfigureWithoutCarrier' on LAN bridge and set 'no-carrier'
as the minimum acceptable operational state for the bridge for
networkd to consider the network 'up'.