Allow hostapd to manage wireless NIC config
With the existing configuration there was a race on boot between
the 2 hostapd instances and networkd to see who would create the
bridge. When one of the hostapd instances wins the race (which
always seemed to be the case) it would create and own the bridge.
This didn't affect runtime function but any stopping/restarting
of that winning hostapd instance would remove the bridge
(since it 'owned' it) and bring the network down.

This change adds a 'wait-online' dependency to both hostapd
instances allowing networkd to always win the race on boot. Both
hostapd instances can now be rebooted independently without
bringing down the network, similar to unplugging an ethernet cable
as one would expect.

A side affect of this change is that another mechanism is needed
to apply any bridge port settings previously handled by networkd.
ExecPostStart handles this, but for this to work, the hostapd
service units needed to change from 'simple' to 'forking. Without
this change the ExecPostStart commands execute before hostapd has
fully started and created the bridge ports.